revise code
